https://youtu.be/rCEf7p3FRTM?t=37 THE WORST YEAR IN PROGRAMMING HISTORY
> I have a running hypothesis that it's generally easier for an university grad to buy an .ai domain and raise VC money than it is to get an entry level job.

Lol. I really picked a bad year to get fired. But actually no, this was a great year to get fired if you want a break. In a few years there will be a flushout of wannabe programmers and I'll have a proper job market then.

I just need to not run out of money by then. Which shouldn't be hard given how large my stash is.

I'll be able to get a job again in 2028 - 2029 if by then I haven't succeeded at either trading or poker botting. Or the venture with Thomas doesn't gain traction.